+++ src/business_logic/transaction/objects/v2/declare_v2.rs
@@ -31,7 +31,7 @@ pub struct InternalDeclareV2 {
     pub tx_type: TransactionType,
     pub validate_entry_point_selector: Felt252,
     pub version: u64,
-    pub max_fee: u64,
+    pub max_fee: u128,
     pub signature: Vec<Felt252>,
     pub nonce: Felt252,
     pub compiled_class_hash: Felt252,
@@ -47,7 +47,7 @@ impl InternalDeclareV2 {
         compiled_class_hash: Felt252,
         chain_id: Felt252,
         sender_address: Address,
-        max_fee: u64,
+        max_fee: u128,
         version: u64,
         signature: Vec<Felt252>,
         nonce: Felt252,
@@ -181,7 +181,7 @@ impl InternalDeclareV2 {
         &mut self,
         state: &mut S,
         general_config: &StarknetGeneralConfig,
-        remaining_gas: u64,
+        remaining_gas: u128,
     ) -> Result<TransactionExecutionInfo, TransactionError> {
         self.verify_version()?;
 
@@ -232,11 +232,11 @@ impl InternalDeclareV2 {
 
     fn run_validate_entrypoint<S: Default + State + StateReader + Clone>(
         &mut self,
-        mut remaining_gas: u64,
+        mut remaining_gas: u128,
         state: &mut S,
         resources_manager: &mut ExecutionResourcesManager,
         general_config: &StarknetGeneralConfig,
-    ) -> Result<(CallInfo, u64), TransactionError> {
+    ) -> Result<(CallInfo, u128), TransactionError> {
         let calldata = [self.compiled_class_hash.clone()].to_vec();
 
         let entry_point = ExecutionEntryPoint {
